+++ package/core/src/components/index.ts
@@ -28,3 +28,6 @@ export { default as Textarea } from "./form/Textarea/Textarea";
 export { default as Switch } from "./form/Switch/Switch";
 export { default as SelectBox } from "./form/SelectBox/SelectBox";
 export { default as MaskInput } from "./form/MaskInput/MaskInput";
+
+// overlays
+export { default as LoadingOverlay } from "./overlays/LoadingOverlay/LoadingOverlay";