+++ package/core/src/contexts/index.ts
@@ -1,3 +1,4 @@
 export { default as ThemeProvider } from "./ThemeProvider";
 export { default as SessionProvider } from "./SessionProvider";
 export { default as ColorSchemeProvider } from "./ColorSchemeProvider";
+export { default as LoadingOverlayProvider } from "./LoadingOverlayProvider";