+++ package/core/src/helpers/index.ts
@@ -1,3 +1,4 @@
 export { default as GetColorValue } from "./GetColorValue";
 export { default as GetTextSizeValue } from "./GetTextSizeValue";
 export { default as GetHeadingSizeValue } from "./GetHeadingSizeValue";
+export { default as HexToRGBA } from "./HexToRGBA";