+++ package/core/src/hooks/index.ts
@@ -3,3 +3,4 @@ export { default as useKeyboard } from "./useKeyboard";
 export { default as useSession } from "./useSession";
 export { default as useTheme } from "./useTheme";
 export { default as useColor } from "./useColor";
+export { default as useLoadingOverlay } from "./useLoadingOverlay";