+++ package/core/src/hooks/useLoadingOverlay.tsx
@@ -1,0 +1,14 @@
+import { useContext } from "react";
+import {
+  LoadingOverlayContext,
+  LoadingOverlayContextDispatch,
+} from "../contexts/LoadingOverlayProvider";
+
+const useLoadingOverlay = () => {
+  const isLoading = useContext(LoadingOverlayContext);
+  const setLoading = useContext(LoadingOverlayContextDispatch);
+
+  return { isLoading, setLoading };
+};
+
+export default useLoadingOverlay;